requirements.txt
requests
bs4
discord.py
pandas
tabula-py
